chore: add functionality to generate BLS keys
