+++ package.json
@@ -9,7 +9,6 @@
         "message-verify": "node dist/actions/message-verify.js",
         "wallet": "node dist/actions/wallet-data.js",
         "validator": "node dist/actions/validator-data.js",
-        "bls-keys": "node dist/actions/validator-data.js --locales",
         "fixtures": "node dist/actions/generate-fixtures.js",
         "get-nonce": "node dist/actions/get-nonce.js",
         "build-contract": "node dist/actions/build-contract.js",
