+++ pnpm-workspace.yaml
@@ -1,9 +1,18 @@
 allowBuilds:
-    bcrypto: true
-    better-sqlite3: true
-    bstring: true
-    es5-ext: true
-    lmdb: true
-    msgpackr-extract: true
-    nsfw: true
-    protobufjs: true
+  bcrypto: true
+  better-sqlite3: true
+  bstring: true
+  es5-ext: true
+  lmdb: true
+  msgpackr-extract: true
+  nsfw: true
+  protobufjs: true
+
+onlyBuiltDependencies:
+  - bcrypto
+  - better-sqlite3
+  - bstring
+  - lmdb
+  - msgpackr-extract
+  - nsfw
+  - protobufjs
